# [fix][java-client] Fix performance regression